a2a_nexus/requirements.txt
fastapi
uvicorn
groq
python-dotenv
jinja2
python-multipart
websockets
